Added More Functionality & Components
Created the complete task modal and hooked it up to event handlers to close/open. Completing a task will now edit the state and remove the task, just like the del handler does. I need to add some sort of functionality in PHP to post the completed tasks to a database,

Also added a new component named TaskStats.js which provides analytics on how tasks are assigned. I felt like the UI needed something to break it up so it wasn't just a long dynamic list.

Also made the app responsive for large tablets and mobiles -at 600px we will break the layout down into a single column and at 1200px we decrease the task info width to keep in line with the smalelr screen sizes.