Worked on Backend
Now have the front end and back end talking to each other. I added an axios call which calls a PHP file on Vidahost. This file queries the database and returns JSON data which is parsed and inputted into my ticket component.

I also added lodash to throttle the requests so we don't slow down the browser or kill the DB. I've set this to make 4 requests a minute which I think is optimal.