Added LoginPage Component
Added the LoginPage component to the app. So far it renders and returns a response but I need to fix axios/php to get the status code from auth.php and return it to the app. Once I have the status code I can make the app derender the LoginPage component once logged in successfully.